fix: move dialog utility to separate file to fix CI #81
